Undo previous move of IO and friends to core
