+++ tools/src/main/scala/scala/scalanative/codegen/CodeGen.scala
@@ -5,13 +5,14 @@ import java.io.File
 import java.nio.file.{Path, Paths, Files}
 import scala.collection.mutable
 import scala.scalanative.build.Config
-import scala.scalanative.build.core.ScalaNative.{dumpDefns, encodedMainClass}
+import scala.scalanative.build.ScalaNative.{dumpDefns, encodedMainClass}
 import scala.scalanative.io.VirtualDirectory
 import scala.scalanative.nir._
 import scala.scalanative.util.{Scope, partitionBy, procs}
 import scala.scalanative.compat.CompatParColls.Converters._
 import java.nio.file.StandardCopyOption
 
+import scala.scalanative.build.ScalaNative
 object CodeGen {
 
   /** Lower and generate code for given assembly. */
