+++ tools/src/test/scala/scala/scalanative/LinkerSpec.scala
@@ -4,7 +4,7 @@ import scala.language.implicitConversions
 import java.io.File
 import java.nio.file.{Files, Path, Paths}
 import scalanative.build.{Config, NativeConfig}
-import scalanative.build.core.ScalaNative
+import scalanative.build.ScalaNative
 import scalanative.util.Scope
 import org.scalatest.flatspec.AnyFlatSpec
