+++ tools/src/test/scala/scala/scalanative/OptimizerSpec.scala
@@ -1,7 +1,6 @@
 package scala.scalanative
 
-import scala.scalanative.build.{Config, NativeConfig, Mode}
-import scala.scalanative.build.core.ScalaNative
+import scala.scalanative.build.{Config, NativeConfig, Mode, ScalaNative}
 
 /** Base class to test the optimizer */
 abstract class OptimizerSpec extends LinkerSpec {
