+++ tools/src/test/scala/scala/scalanative/codegen/CodeGenSpec.scala
@@ -4,7 +4,7 @@ package codegen
 import java.nio.file.{Path, Paths}
 import scalanative.io.VirtualDirectory
 import scalanative.build.Config
-import scalanative.build.core.ScalaNative
+import scalanative.build.ScalaNative
 import scalanative.util.Scope
 
 /** Base class to test code generation */
